[hotfix][connector-kafka] Merge KafkaITCase and KafkaTableTestBase
